Refactor Message
This allows users of library to create their own raw nats::Message
instances without the need to have Client injected
