+++ src/subscription.rs
@@ -18,8 +18,7 @@ use std::time::Duration;
 
 use crossbeam_channel as channel;
 
-use crate::client::Client;
-use crate::message::Message;
+use crate::client::{self, Client};
 
 #[derive(Debug)]
 struct Inner {
@@ -30,7 +29,7 @@ struct Inner {
     pub(crate) subject: String,
 
     /// MSG operations received from the server.
-    pub(crate) messages: channel::Receiver<Message>,
+    pub(crate) messages: channel::Receiver<client::Message>,
 
     /// Client associated with subscription.
     pub(crate) client: Client,
@@ -52,7 +51,7 @@ impl Subscription {
     pub(crate) fn new(
         sid: u64,
         subject: String,
-        messages: channel::Receiver<Message>,
+        messages: channel::Receiver<client::Message>,
         client: Client,
     ) -> Subscription {
         Subscription(Arc::new(Inner {
@@ -87,7 +86,7 @@ impl Subscription {
     /// }
     /// # }
     /// ```
-    pub fn receiver(&self) -> &channel::Receiver<Message> {
+    pub fn receiver(&self) -> &channel::Receiver<client::Message> {
         &self.0.messages
     }
 
@@ -104,7 +103,7 @@ impl Subscription {
     /// # Ok(())
     /// # }
     /// ```
-    pub fn next(&self) -> Option<Message> {
+    pub fn next(&self) -> Option<client::Message> {
         self.0.messages.recv().ok()
     }
 
@@ -123,7 +122,7 @@ impl Subscription {
     /// # Ok(())
     /// # }
     /// ```
-    pub fn try_next(&self) -> Option<Message> {
+    pub fn try_next(&self) -> Option<client::Message> {
         self.0.messages.try_recv().ok()
     }
 
@@ -139,7 +138,7 @@ impl Subscription {
     /// # Ok(())
     /// # }
     /// ```
-    pub fn next_timeout(&self, timeout: Duration) -> io::Result<Message> {
+    pub fn next_timeout(&self, timeout: Duration) -> io::Result<client::Message> {
         match self.0.messages.recv_timeout(timeout) {
             Ok(msg) => Ok(msg),
             Err(channel::RecvTimeoutError::Timeout) => Err(io::Error::new(
@@ -237,7 +236,7 @@ impl Subscription {
     /// ```
     pub fn with_handler<F>(self, handler: F) -> Handler
     where
-        F: Fn(Message) -> io::Result<()> + Send + 'static,
+        F: Fn(client::Message) -> io::Result<()> + Send + 'static,
     {
         // This will allow us to not have to capture the return. When it is
         // dropped it will not unsubscribe from the server.
@@ -340,7 +339,7 @@ impl Subscription {
 }
 
 impl IntoIterator for Subscription {
-    type Item = Message;
+    type Item = client::Message;
     type IntoIter = IntoIter;
 
     fn into_iter(self) -> IntoIter {
@@ -349,7 +348,7 @@ impl IntoIterator for Subscription {
 }
 
 impl<'a> IntoIterator for &'a Subscription {
-    type Item = Message;
+    type Item = client::Message;
     type IntoIter = Iter<'a>;
 
     fn into_iter(self) -> Iter<'a> {
@@ -388,7 +387,7 @@ pub struct TryIter<'a> {
 }
 
 impl<'a> Iterator for TryIter<'a> {
-    type Item = Message;
+    type Item = client::Message;
     fn next(&mut self) -> Option<Self::Item> {
         self.subscription.try_next()
     }
@@ -400,7 +399,7 @@ pub struct Iter<'a> {
 }
 
 impl<'a> Iterator for Iter<'a> {
-    type Item = Message;
+    type Item = client::Message;
     fn next(&mut self) -> Option<Self::Item> {
         self.subscription.next()
     }
@@ -412,7 +411,7 @@ pub struct IntoIter {
 }
 
 impl Iterator for IntoIter {
-    type Item = Message;
+    type Item = client::Message;
     fn next(&mut self) -> Option<Self::Item> {
         self.subscription.next()
     }
@@ -427,7 +426,7 @@ pub struct TimeoutIter<'a> {
 }
 
 impl<'a> Iterator for TimeoutIter<'a> {
-    type Item = Message;
+    type Item = client::Message;
     fn next(&mut self) -> Option<Self::Item> {
         self.subscription.next_timeout(self.to).ok()
     }
